telemetry/obs-signal-ingestor/requirements.txt
kubernetes==29.0.0
pyyaml==6.0.2
requests==2.32.3
